Transformation-Based Ray Tracing
Each object contains a tranformation matrix (and its inverse) which is used to transform a ray into the local space of the object. Ray intersection with primitives can be greatly simplified if certain properties can be fixed. For example, a unit sphere centered at the origin is easier to ray trace than an oblate spheroid arbitrarily placed and rotated in the scene. This technique will be further expanded upon when implementing relativistic effects, where the camera rays will need to be transformed between inertial reference frames.

Other changes:
A Hit struct is updated by the raycasting functions, which stores information about the surface that was hit.
Translation, Rotation, and Scaling matrix is generated by a call to TRS().