First-Layer Octree Intersection
A single-element octree is generated to act as an Axis-Aligned Bounding Box (AABB). This AABB is used in the shader to do a preliminary check for whether or not a ray could potentially intersect the mesh. This excludes any region of the screen that isn't covered by the AABB, significantly improving performance over the previous version when the object does not extend over the entire screen.