Basic Octree Traversal
Iterates through octree in a pre-order traversal. Any cuboids that are not intersected by the ray are ignored. Every leaf cuboid is checked for triangle intersections.
One performance improvement will be to modify the order in which cuboids are iterated. If they are iterated in the order the ray intersects them, the closest hit is more likely to be found early.