Bug Fixes and Incomplete Ordered Traversal
Fixed a bug involving the order an octree cuboid's children are stored. The getChildIndex() function in the kernel appeared to be returning an incorrect value, but in actuality the index of the child was correct according to the expected layout, but the children were not in their correct locations.

Working on implementing ordered traversal of the octree, where closer cuboids are visited first. Algorithm used to move from one cuboid to its neighbor seems to be producing incorrect results often. Possibly incorrect neighbor index calculation as well?