Octree (#1)
* First-Layer Octree Intersection

A single-element octree is generated to act as an Axis-Aligned Bounding Box (AABB). This AABB is used in the shader to do a preliminary check for whether or not a ray could potentially intersect the mesh. This excludes any region of the screen that isn't covered by the AABB, significantly improving performance over the previous version when the object does not extend over the entire screen.

* Octree Subdivision - Child Creation

Octree cuboids can be subdivided into 8 children, and their neighbor indices are correctly calculated. Triangle overlap is not yet computed.

* Incomplete Octree Intersection

* Basic Octree Traversal

Iterates through octree in a pre-order traversal. Any cuboids that are not intersected by the ray are ignored. Every leaf cuboid is checked for triangle intersections.
One performance improvement will be to modify the order in which cuboids are iterated. If they are iterated in the order the ray intersects them, the closest hit is more likely to be found early.

* Bug Fixes and Incomplete Ordered Traversal

Fixed a bug involving the order an octree cuboid's children are stored. The getChildIndex() function in the kernel appeared to be returning an incorrect value, but in actuality the index of the child was correct according to the expected layout, but the children were not in their correct locations.

Working on implementing ordered traversal of the octree, where closer cuboids are visited first. Algorithm used to move from one cuboid to its neighbor seems to be producing incorrect results often. Possibly incorrect neighbor index calculation as well?

* Fixed Neighbor Calculation

Previously neighbors were hard-coded. Now they are computed algorithmically, which fixes an error that existed in the hard-coded values. Octree traversal now works as expected.

* Fixed Triangle Bounds +Added Lighting + Octree JSON Representation

-Fixed a bug involving the octree-triangle intersection test. When subdividing the octree, each octants bounds were used when half-bounds should have been used instead. Now, exactly the triangles overlapped by each octant are added to its list of triangles.
-Added a function for generating a .json file representation of the octree. The json structure includes a field for the Mathematica code for visualizing an octant's bounds and every triangle contained within.
-Added basic lighting to the kernel: A fixed-position point light source is used for calculating direct lighting on all surfaces, with an additional term for ambient lighting. Shadows are not yet considered, but should be fairly simple to incorporate into the ray tracing step.