Multiple Meshes, Improved Normal Generation, Shadow Casting and Reflection
Multiple meshes can now be imported into a scene.
The Mesh struct contains a vector called meshIndices which gives the index in the octree list for the root node of each mesh.
The Object struct contains a new field called meshIndex, which needs to be set to the value at Mesh.meshIndices[n] to give the object the nth mesh.
The objectType enum now gives the option to specify an object is a MESH type, and the kernel will ray trace it accordingly.

If an OBJ file does not provide normals, they can be calculated when the mesh is imported. Previously, each triangle would have one normal associated with all 3 of its vertices calculated as the normal to the triangle's face. Now, each triangle associated with each vector is collected, and the weighted sum of their normals are calculated to be associated with that vector. The weights are proportional to the area of the triangle, so larger tris will have a greater impact on their vertices' normals. Previously, computed normals would give the mesh a faceted look, because the normal would be constant across the surface of each triangle. Now, each vertex has a unique normal so the surface is given a smooth appearance.

Shadow casting was slightly improved in the kernel. Rays are now send from the light towards the object, and the hit object's ID is compared against the expected ID to determine shadowing.
A bug in the sphere ray casting algorithm caused spheres behind the ray start point to be hit, so now spheres behind the camera won't be rendered and won't cast shadows forwards towards the light.
Meshes now enable reflection in the kernel. When an object of MESH type is hit, a secondary ray is reflected over the normal and used to give the object a mirrored appearance. This was primarily added to test the normal generation algorithm, and might not be kept in the final version.