Texture Mapping
Textures can be imported using CImg, which natively supports RAW, ASC, HDR, INR, PPM, PGM, BMP, PAN, and DLM image files, and JPG, GIF, PNG, and TIF if ImageMagick is installed on the user's system. Objects have 3 new fields, for texture index in the pixel array, texture width, and texture height. If the texture index is not -1, the kernel will use the object's UV coordinates (either provided in the mesh or determined by the primitive type) to sample the texture at that index.