Kernel Performance Improvements
Fixed a bug in the case when a ray originates within a mesh's octree. Also, improved performance when casting a ray from within an octree facing outwards. Now, very little of the octree needs to be checked.
Modified octree subdivision routine to calculate the maximum number of triangles sharing a common vertex. This value is used to cut off subdivision when an octant contains that number of triangles.
Modified framerate calculation to reevaluate every frame, as well as display overall average framerate.
getChildIndex() in the kernel replaced with one-liner equivalent.
sample_light() added to kernel: Alternative to intersect_scene() which seeks to minimize calculations when just determining intersection, without color or normal information. Returns on first intersection found, used only to determine shadows.