Split into Multiple Files + Flashing + Screenshots
The primary file, main.cpp, has been divided between several different header and source files.

The Object struct and kernel have been updated to support periodic flashing. Objects can be given a period and duration (where 0 < duration < period) which will cause them to increase in brightness for that duration in seconds once every period. This allows the observer to better visualize the effects of time dilation, as moving objects will appear to flash at a lower frequency. Relativity of simultaneity can also be seen, as a simultaneous flash across the surface of an object will not be simultaneous if observed from a relatively moving frame.
Screenshots and videos added to demonstrate some common scenes.

A scene demonstrating the Ladder Paradox was created.